requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
azure-core==1.29.5
azure-storage-blob==12.19.0
beanie==1.23.6
httpcore==1.0.2
httpx==0.25.2
isodate==0.6.1
json-repair==0.4.3
lazy-model==0.2.0
motor==3.3.2
multidict==6.0.4
openai==1.3.7
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pymongo==4.6.1
python-dotenv==1.0.0
python-magic==0.4.27
regex==2023.10.3
requests==2.31.0
tiktoken==0.5.2
urllib3==2.1.0
srt==3.5.3